Restored index_list to a 2D list.
The index_list now has a list of block_lists[].
This means each element in the Index_list represents a row. if you access that row you will find a list of block objects representing the "set" in that row.
Basically to access index 12 block 3 and update the tag you do the following ==> 
index_list[11][2].tag = "some tag"