Fixed small cache errors. index out of range error. see description.
The calculated values were truncating the decimal points with "//" and math.ceil. These decimals are needed when the number of rows and blocks are below a kilobyte. Essentially if the number of rows were 512, that would be .5 kb. These truncations set them to 0 which cascaded errors through the cache simulation.