Implemented RR and RND replacement algorithms
Algorithm will only run when the set is full. That is: No tag matches and no empty blocks left to replace.

RR replacement algorithm. It sets a "next" flag inside the Block object. It searches for the object with that flag set true, replaces it and then updates the next block's flag to true. 

RND - Literally grabs a random number between 0 and the associativity of the cache. That number then selects the block in the set to be replaced.